Consolidate text input editors under Editors/text/
Move InputText/InputTextArea editors and TableInputText/TableInputTextArea/
StringList variants into Editors/text/.
