+++ src/main/cliapp/src/components/Editors/text/InputTextTableEditor.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { InputTextEditor } from '../../InputTextEditor';
+import { InputTextEditor } from './InputTextEditor';
 import { ErrorMessageComponent } from '../../Error/ErrorMessageComponent';
 
 export const InputTextTableEditor = ({ editorOptions, field, errorMessagesRef }) => {
