+++ src/main/cliapp/src/components/Editors/text/StringListTableEditor.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StringListEditor } from '../StringListEditor';
+import { StringListEditor } from './StringListEditor';
 import { ErrorMessageComponent } from '../../Error/ErrorMessageComponent';
 
 export const StringListTableEditor = ({ editorOptions, field, errorMessagesRef }) => {
