+++ src/main/cliapp/src/containers/allelesPage/functionalImpacts/FunctionalImpactsFormTable.jsx
@@ -7,7 +7,7 @@ import { InternalEditor } from '../../../components/Editors/InternalEditor';
 import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
 import { FunctionalImpactsEditor } from '../../../components/Editors/FunctionalImpactsEditor';
 import { PhenotypeTermEditor } from '../../../components/Editors/PhenotypeTermEditor';
-import { TableInputTextEditor } from '../../../components/Editors/TableInputTextEditor';
+import { TableInputTextEditor } from '../../../components/Editors/text/TableInputTextEditor';
 
 export const FunctionalImpactsFormTable = ({
 	functionalImpacts,
