+++ src/main/cliapp/src/containers/allelesPage/inheritanceModes/InheritanceModesEditDialog.jsx
@@ -12,7 +12,7 @@ import { InternalEditor } from '../../../components/Editors/InternalEditor';
 import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
 import { ControlledVocabularyEditor } from '../../../components/Editors/ControlledVocabularyEditor';
 import { PhenotypeTermEditor } from '../../../components/Editors/PhenotypeTermEditor';
-import { TableInputTextAreaEditor } from '../../../components/Editors/TableInputTextAreaEditor';
+import { TableInputTextAreaEditor } from '../../../components/Editors/text/TableInputTextAreaEditor';
 import { Endpoints } from '../../../constants/Endpoints';
 import { defaultAutocompleteOnChange } from '../../../utils/utils';
